Actually fix Windows support
The script now properly runs on Windows, through a wrapper I will PR
separately.

To make this work, I've had to not strip backslashes from the input (by
adding -r to the read calls). They are being specified as path
separators, and stripping them screws up paths.

I also swapped out the right-trim for a more specific
remove-trailing-CR-if-present, which is now uniformly applied to both
instances of read.

I also left in my check for trying to upload a nonexistent file (which
happened when backslashes were stripped). It *should* never happen, but
it's definitely an error if it does.