unglobber the error stream
When checking if a file is present, the size function of rclone clobbers the output as described in issue #37.
That patch stores the error stream and the grepped result in variables.
These variables are printed with prefix DEBUG to stdout. That has the advantage, that git annex with debug flag, shows these messages.