RF+ENH(TST): move all testing into a script which could be ran locally as well
It also extends testing with actually getting the file.  Original file of 0 length getting
does not work for me at all! See comment in the code.  I left it in, and if fails on CI
better be disabled until fixed. For file of size 1 -- works ok, and works ok with the trickier
tests later on (again non-0 size)